bot/requirements.txt
# Telegram
aiogram==3.13.1
aiohttp>=3.9.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0

# HTTP / API
httpx>=0.27.0
aiohttp>=3.9.0

# Utils
python-dotenv>=1.0.0
